+++ src/types/resource.ts
@@ -24,3 +24,48 @@ export interface Resource {
    */
   url: string;
 }
+
+/**
+ * Partial Item Config
+ */
+export interface PartialItemConfig {
+  /**
+   * Id
+   *
+   * @type {string}
+   */
+  id: string;
+
+  /**
+   * Url
+   *
+   * @type {string}
+   */
+  url: string;
+
+  /**
+   * Name
+   *
+   * @type {string}
+   */
+  name: string;
+}
+
+/**
+ * Partial Item
+ */
+export interface PartialItem {
+  /**
+   * Content
+   *
+   * @type {string}
+   */
+  content: string;
+
+  /**
+   * Name
+   *
+   * @type {string}
+   */
+  name: string;
+}