+++ src/utils/rows.ts
@@ -1,0 +1,13 @@
+/**
+ * Reorder
+ * @param list
+ * @param startIndex
+ * @param endIndex
+ */
+export const reorder = <T>(list: T[], startIndex: number, endIndex: number) => {
+  const result = Array.from(list);
+  const [removed] = result.splice(startIndex, 1);
+  result.splice(endIndex, 0, removed);
+
+  return result;
+};