+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / Enhancements
 
 - Updated Provisioning Dashboard (#346)
+- Added partials to load external HTML (#345)
 
 ## 5.2.0 (2024-07-25)